pyproject.toml
[project]
name = "langgraph-teams-bot"
version = "0.0.0"
description = "LangGraph Teams Bot"
authors = [{ name = "Microsoft", email = "teams@microsoft.com"}]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "teams-ai (>=1.8.1,<2.0.0)",
    "langchain (>=0.2.10,<1.0.0)",
    "langchain-openai (>=0.1.8,<1.0.0)",
    "langgraph (>=0.2.0,<1.0.0)"
]

[tool.poetry]
packages = [
    { include = "src" }
]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pylint = "^2.17.4"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
black = "^24.3.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[project.scripts]
lint = "scripts:lint"
fmt = "scripts:fmt"
test = "scripts:test"
clean = "scripts:clean"
ci = "scripts:ci"
start = "scripts:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = "--cov-report html:coverage --cov=src"

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
show_error_codes = true
no_implicit_optional = true
warn_unused_ignores = true